server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
openai>=1.0.0
